Move common commands to `commoncommands` (#2660)
Refs #2650.
