+++ internal/handlers/commoncommands/commoncommands.go
@@ -12,22 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package common
-
-import (
-	"context"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	"github.com/FerretDB/FerretDB/internal/wire"
-)
-
-// TestMsgWhatsMyURI checks a special case: if context is not set, it panics.
-// The "normal" cases are covered in integration tests for MsgWhatsMyURI command.
-func TestMsgWhatsMyURI(t *testing.T) {
-	require.Panics(t, func() {
-		_, err := MsgWhatsMyURI(context.Background(), new(wire.OpMsg))
-		require.NoError(t, err)
-	})
-}
+// Package commoncommands provides command handlers shared by all handlers.
+package commoncommands
