+++ internal/handlers/commoncommands/msg_getcmdlineopts.go
@@ -12,23 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package common
+package commoncommands
 
 import (
 	"context"
 
-	"github.com/FerretDB/FerretDB/internal/clientconn/conninfo"
 	"github.com/FerretDB/FerretDB/internal/types"
 	"github.com/FerretDB/FerretDB/internal/util/must"
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgWhatsMyURI is a common implementation of the whatsMyURI command.
-func MsgWhatsMyURI(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
+// MsgGetCmdLineOpts is a common implementation of the getCmdLineOpts command.
+func MsgGetCmdLineOpts(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	var reply wire.OpMsg
 	must.NoError(reply.SetSections(wire.OpMsgSection{
 		Documents: []*types.Document{must.NotFail(types.NewDocument(
-			"you", conninfo.Get(ctx).PeerAddr,
+			"argv", must.NotFail(types.NewArray("ferretdb")),
+			"parsed", must.NotFail(types.NewDocument()),
 			"ok", float64(1),
 		))},
 	}))
