+++ internal/handlers/pg/msg_whatsmyuri.go
@@ -17,11 +17,11 @@ package pg
 import (
 	"context"
 
-	"github.com/FerretDB/FerretDB/internal/handlers/common"
+	"github.com/FerretDB/FerretDB/internal/handlers/commoncommands"
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
 // MsgWhatsMyURI implements HandlerInterface.
 func (h *Handler) MsgWhatsMyURI(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
-	return common.MsgWhatsMyURI(ctx, msg)
+	return commoncommands.MsgWhatsMyURI(ctx, msg)
 }
